refactor: relocate artifact-ref types into homeboy-artifact-ref-contract off observation
The artifact_ref module was pure serde data (ArtifactReference enum,
ArtifactRef/EvidenceRef structs, reviewer-facing validation, artifact_uri)
EXCEPT for two conversion methods that took &observation::ArtifactRecord
-- the sole edge coupling it to core's observation layer.

Split it: the pure types move DOWN into a new leaf crate
homeboy-artifact-ref-contract (deps: serde + engine-primitives codecs
only). The two ArtifactRecord conversions become free functions in core
(artifact_ref_from_record / evidence_ref_for_artifact), mirroring the
existing artifact_contract pattern where observation-coupled conversions
stay in core. Core re-exports the contract as artifact_ref, so all
existing paths (core-internal + cli + fuzz) resolve unchanged; the 2
conversion call sites (core evidence_report, fuzz) repoint to the free fns.

This relocates ArtifactReference off its observation dependency, which is
the prerequisite that unblocks a future full execution_contract
extraction (execution_contract builds on ArtifactReference::parse).
