+++ crates/contracts/homeboy-artifact-ref-contract/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "homeboy-artifact-ref-contract"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+authors = ["Chris Huber <chubes@extrachill.com>"]
+description = "Pure serializable artifact-reference contract types for homeboy"
+repository = "https://github.com/Extra-Chill/homeboy"
+publish = false
+
+[lib]
+name = "homeboy_artifact_ref_contract"
+path = "src/lib.rs"
+
+[dependencies]
+homeboy-engine-primitives = { version = "0.1.0", path = "../../homeboy-engine-primitives" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
